move generic functions to plot_helpers
